requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
airflow-provider-great-expectations==0.2.6
apache-airflow-providers-common-sql==1.5.2
pymssql==2.1.5
pyodbc==4.0.39
apache-airflow-providers-microsoft-mssql==3.4.1


